Initialize constructors and some classes.
This commit initializes some of the constructors that we need.

To verify a word's existence, my plan is to use wordsapi. If the
response we get from the API for the inputted word is 404, that
would mean that the word actually does not exist. Else, it would,
we can assume that the word exists. This implementation for the
API will be done using a different class altogether which I have
initialized.